add variable environment
